# openapi/sis-api.yaml
openapi: 3.0.1
info:
  title: Smart Insurance System
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /user/{user_id}/profile:
    get:
      tags:
        - User Profile
      summary: Get a user profile
      operationId: get_profile
      parameters:
        - name: user_id
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successfully retrived user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        404:
          description: User doesn't exist
          content: { }
    post:
      tags:
        - User Profile
      summary: Create a user profile
      operationId: create_profile
      parameters:
        - name: user_id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
        required: true
      responses:
        200:
          description: Successfully updated user
          content: { }
        404:
          description: User doesn't exist
          content: { }
      x-codegen-request-body-name: UserProfile
    put:
      tags:
        - User Profile
      summary: Update and replace a user profile
      operationId: update_profile
      parameters:
        - name: user_id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
        required: true
      responses:
        200:
          description: Successfully updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        404:
          description: User doesn't exist
          content: { }
      x-codegen-request-body-name: User
    delete:
      tags:
        - User Profile
      summary: Remove a user profile
      operationId: remove_profile
      parameters:
        - name: user_id
          in: path
          description: The id of the user to remove
          required: true
          schema:
            type: integer
      responses:
        202:
          description: Successfully deleted user
          content: { }
        404:
          description: User doesn't exist
          content: { }
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: X-Auth
      in: header
      x-apikeyInfoFunc: check_api_key
  schemas:
...
    UserProfile:
      type: object
      properties:
        user_id:
          type: integer
        personal_details:
          $ref: '#/components/schemas/PersonalDetails'
        driver_history:
          $ref: '#/components/schemas/DriverHistory'
    PersonalDetails:
      properties:
        id:
          type: integer
        full_name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        dob:
          type: string
          format: date
        relationship_status:
          type: string
          nullable: true
          enum:
            - Single
            - Married
            - null
        home_owner:
          type: boolean
        dependents:
          description: Number of dependents under the age of 16
          type: integer
        employment_status:
          type: string
          nullable: true
          enum:
            - FullTime
            - PartTime
            - Unemployed
            - Retired
            - Student
            - null
...
    DriverHistory:
      properties:
        id:
          type: integer
        licence_type:
          type: string
          enum:
            - Full
            - Provisional
            - null
        license_since:
          type: string
          format: date
        licence_no:
          type: string
        previous_claims:
          type: array
          items:
            $ref: '#/components/schemas/PreviousClaim'
    PreviousClaim:
      properties:
        id:
          type: integer
        driver_history_id:
          type: integer
        date:
          type: string
          format: date
        fault:
          type: string
          enum:
            - Self
            - ThirdParty
        claim_type:
          type: string
          enum:
            - Accident
            - Theft
...
    Address:
      properties:
        id:
          type: integer
        number_or_name:
          type: string
        street:
          type: string
        town:
          type: string
        county:
          type: string
        postcode:
          type: string
...