# sis-api/swagger_server/models/user_profile.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.driver_history import DriverHistory  # noqa: F401,E501
from swagger_server.models.personal_details import PersonalDetails  # noqa: F401,E501
from swagger_server import util


class UserProfile(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, user_id: int=None, personal_details: PersonalDetails=None, driver_history: DriverHistory=None):  # noqa: E501
        """UserProfile - a model defined in Swagger

        :param user_id: The user_id of this UserProfile.  # noqa: E501
        :type user_id: int
        :param personal_details: The personal_details of this UserProfile.  # noqa: E501
        :type personal_details: PersonalDetails
        :param driver_history: The driver_history of this UserProfile.  # noqa: E501
        :type driver_history: DriverHistory
        """
        self.swagger_types = {
            'user_id': int,
            'personal_details': PersonalDetails,
            'driver_history': DriverHistory
        }

        self.attribute_map = {
            'user_id': 'user_id',
            'personal_details': 'personal_details',
            'driver_history': 'driver_history'
        }
        self._user_id = user_id
        self._personal_details = personal_details
        self._driver_history = driver_history

    @classmethod
    def from_dict(cls, dikt) -> 'UserProfile':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UserProfile of this UserProfile.  # noqa: E501
        :rtype: UserProfile
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self) -> int:
        """Gets the user_id of this UserProfile.


        :return: The user_id of this UserProfile.
        :rtype: int
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: int):
        """Sets the user_id of this UserProfile.


        :param user_id: The user_id of this UserProfile.
        :type user_id: int
        """

        self._user_id = user_id

    @property
    def personal_details(self) -> PersonalDetails:
        """Gets the personal_details of this UserProfile.


        :return: The personal_details of this UserProfile.
        :rtype: PersonalDetails
        """
        return self._personal_details

    @personal_details.setter
    def personal_details(self, personal_details: PersonalDetails):
        """Sets the personal_details of this UserProfile.


        :param personal_details: The personal_details of this UserProfile.
        :type personal_details: PersonalDetails
        """

        self._personal_details = personal_details

    @property
    def driver_history(self) -> DriverHistory:
        """Gets the driver_history of this UserProfile.


        :return: The driver_history of this UserProfile.
        :rtype: DriverHistory
        """
        return self._driver_history

    @driver_history.setter
    def driver_history(self, driver_history: DriverHistory):
        """Sets the driver_history of this UserProfile.


        :param driver_history: The driver_history of this UserProfile.
        :type driver_history: DriverHistory
        """

        self._driver_history = driver_history